車が1台走るようにになった

+++ coordinate.cpp
@@ -1,16 +1,16 @@
 class Coordinate{
   private:
-    double x;
-    double y;
+    float x;
+    float y;
 
   public:
-    Coordinate(double x, double y);
-    double get_x(){ return x; }
-    double get_y(){ return y; }
-    void set_x(double new_x){ x = new_x; }
-    void set_y(double new_y){ y = new_y; }
+    Coordinate(float x, float y);
+    float get_x(){ return x; }
+    float get_y(){ return y; }
+    void set_x(float new_x){ x = new_x; }
+    void set_y(float new_y){ y = new_y; }
 };
 
-Coordinate::Coordinate(double init_x, double init_y) : x(init_x), y(init_y)
+Coordinate::Coordinate(float init_x, float init_y) : x(init_x), y(init_y)
 {
 }
